% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboMgITC.competative.R
\name{MetaboMgITC.compete}
\alias{MetaboMgITC.compete}
\title{Generalized ITC data fitting function for competative ITC}
\usage{
MetaboMgITC.compete(
  cell,
  competator,
  blank,
  blank.competator,
  Thermodynamic.equation = "Wiseman.isotherm",
  Fit.start = list(V0 = 1.4247, H = 2534, K = 2880, Hc = 2534, Kc = 2880, c = 0, h = 0,
    Mg.contaminate = 1),
  Remove.injection = 1,
  Saturation.threshold = FALSE,
  Save.path.and.prefix = FALSE
)
}
\arguments{
\item{cell}{Object created by "read.itc" for the ITC run containing the macromolecule.}

\item{competator}{Objecrt created by "read.itc" for the ITC run containing the macromolecule and the competative ligand.}

\item{blank}{Object created by "read.itc" for the ITC run containing the ligand tirated into buffer.}

\item{Thermodynamic.equation}{The thermodynamic model that you want to fit the data to. Default = "wiseman.isotherm".}

\item{Fit.start}{Starting parameters for the non-linear regression. Default = = list(H = 2534, K = 2880).}

\item{Remove.injection}{Injections you want to remove from subsequent analysis. Default = 1 is the standard for ITC experiments. More than one injection can be supplied in a vector.}

\item{Saturation.threshold}{For low c value ITC experiments fit to Wiseman isotherms, macomolecule saturation ranging from 70\% to 90\% produce the same answers. Thus, sometimes it is desirable to standardize the saturation threshold between experiments to minimize degredation of a ligand or macromolecule (for example with ATP binding Mg). Default = FALSE. If set to 0.8, this function will only fit data required to reach 80\% macromolecule saturation.}

\item{competator.blank}{Object created by "read.itc" for the ITC run containing the ligand tirated into buffer with the competator.}
}
\value{
A list containing summary statistics for the fit and a graphical depiction of the fits, the raw ITC curve, and the summary statistics.
}
\description{
Fits an ITC experiment to a user specified thermodynamic model to determine thermodynamic statistics. Raw ".itc" formatted data
should be read in using "read.itc". Fits the data to a thermodynamic model then prints a graphical summary of the fit.
}
